Start to split test.cpp. Move flexbuffers tests
